# Use multiple user contexts in atime_updater benchmark