With this change, https://github.com/buildbuddy-io/buildbuddy/pull/9809 makes the benchmark run about 16% faster:

```
goos: linux
goarch: amd64
cpu: AMD Ryzen 7 5800X 8-Core Processor             
           │ /home/iain/old.txt │         /home/iain/new.txt          │
           │       sec/op       │   sec/op     vs base                │
Enqueue-16          244.5m ± 0%   205.4m ± 1%  -15.99% (p=0.000 n=20)

           │ /home/iain/old.txt │          /home/iain/new.txt          │
           │        B/op        │     B/op      vs base                │
Enqueue-16         196.8Mi ± 0%   263.7Mi ± 0%  +34.01% (p=0.000 n=20)

           │ /home/iain/old.txt │         /home/iain/new.txt          │
           │     allocs/op      │  allocs/op   vs base                │
Enqueue-16          5.130M ± 0%   6.030M ± 0%  +17.54% (p=0.000 n=20)
```